pyproject.toml
[project]
name = "openai-web-summerizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "ipython>=9.8.0",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
